Capstone Project
capstone project in Coursera refers to a final project at the end of a study program. In the real world, these types of projects are more often referred to as case studies, Case studies are common ways for employers to assess the skills of prospective job candidates. In this part of the course, you will explore the benefits of using capstone projects, case studies, and portfolios to showcase your new skills to potential employers.